move out coercion-related stuff into TypeCoercions
