+++ src/main/java/me/hugeblank/allium/lua/type/InvalidArgumentException.java
@@ -1,0 +1,19 @@
+package me.hugeblank.allium.lua.type;
+
+public final class InvalidArgumentException extends Exception {
+    public InvalidArgumentException() {
+        super();
+    }
+
+    public InvalidArgumentException(String message) {
+        super(message);
+    }
+
+    public InvalidArgumentException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public InvalidArgumentException(Throwable cause) {
+        super(cause);
+    }
+}
