+++ src/main/java/me/hugeblank/allium/lua/type/LuaToJavaConverter.java
@@ -4,6 +4,6 @@ import org.squiddev.cobalt.LuaError;
 import org.squiddev.cobalt.LuaState;
 import org.squiddev.cobalt.LuaValue;
 
-public interface LuaDeserializer<T> {
-    T fromLua(LuaState state, LuaValue value) throws LuaError, UserdataFactory.InvalidArgumentException;
+public interface LuaToJavaConverter<T> {
+    T fromLua(LuaState state, LuaValue value) throws LuaError, InvalidArgumentException;
 }
